# public/openapi.yaml
openapi: 3.1.0
info:
  title: ChronoChimp API
  version: v1.0.0
  description: API for managing tasks and user settings in the ChronoChimp application.
servers:
  - url: /api
    description: Main API server
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized: Invalid or missing authentication token"
              error:
                type: string
                example: "unauthorized"
              statusCode:
                type: integer
                example: 401
    ForbiddenError:
      description: Forbidden - Insufficient permissions.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Forbidden"
...
    ValidationError:
      description: Bad Request - Invalid input data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal Server Error - E.g., JWT_SECRET not configured.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal Server Error"
  schemas:
...
    UserRole:
      type: string
      enum: [Admin, Editor, Viewer]
      example: Viewer
    User: # For general user display, not for sensitive creation/update
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "user-abc-123"
        name:
          type: string
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane.doe@example.com"
        role:
          $ref: '#/components/schemas/UserRole'
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: "https://placehold.co/100x100.png"
        isTwoFactorEnabled:
          type: boolean
          description: "Indicates if 2FA is enabled for the user."
          example: false
        emailNotificationsEnabled:
          type: boolean
          description: "User preference for email notifications."
          example: true
        inAppNotificationsEnabled:
          type: boolean
          description: "User preference for in-app notifications."
          example: true
        smtpHost:
          type: string
          nullable: true
          example: "smtp.example.com"
        smtpPort:
          type: integer
          nullable: true
          example: 587
        smtpEncryption:
          type: string
          enum: [none, ssl, starttls]
          nullable: true
          example: "starttls"
        smtpUsername:
          type: string
          nullable: true
          example: "user@example.com"
        smtpSendFrom:
          type: string
          format: email
          nullable: true
          description: "Default 'send from' email for SMTP."
          example: "noreply@example.com"
        joinedDate:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
...
    UserCreationRequest: # Admin creating a user
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: "New User"
        email:
          type: string
          format: email
          example: "new.user@example.com"
        password:
          type: string
          format: password
          minLength: 6
          description: "Password for the new user. Will be hashed by the server."
        role:
          $ref: '#/components/schemas/UserRole'
          default: "Viewer"
        avatarUrl:
          type: string
          format: uri # Or allow data URI
          nullable: true
        isTwoFactorEnabled:
          type: boolean
          default: false
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid input."
        details: # For Zod validation errors
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address."]
            password: ["Password must be at least 6 characters."]
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: "Operation successful."
  securitySchemes:
    bearerAuth: # Can be used to denote JWT authentication
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /auth/login:
    post:
      summary: Log in a user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
                password:
                  type: string
                  format: password
                  example: "sEcReTpAsSwOrD"
      responses:
        '200':
          description: Login successful or 2FA required
          content:
            application/json:
              schema:
                oneOf:
                  - type: object # Normal login
                    properties:
                      user:
                        $ref: '#/components/schemas/User' # Excludes password
                      message:
                        type: string
                        example: "Login successful"
                  - type: object # 2FA Required
                    properties:
                      twoFactorRequired:
                        type: boolean
                        example: true
                      userId:
                        type: string
                        format: uuid
                        example: "user-abc-123"
                      message:
                        type: string
                        example: "Two-Factor Authentication required."
        '400':
          description: Invalid input (e.g., missing email or password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials (email or password incorrect)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid email or password."
        '500':
          $ref: '#/components/responses/ServerError' # For JWT_SECRET config error
...
  /auth/forgot-password:
    post:
      summary: Request a password reset
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        '200':
          description: Password reset request processed (generic message for security)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
                properties:
                  message:
                    example: "If an account with that email exists, instructions to reset your password have been sent."
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /auth/2fa-verify: # For verifying OTP during login
    post:
      summary: Verify 2FA OTP during login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, otp]
              properties:
                userId:
                  type: string
                  format: uuid
                  example: "user-abc-123"
                otp:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: "123456"
      responses:
        '200':
          description: 2FA verification successful, login complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User' # Excludes password/secrets
                  message:
                    type: string
                    example: "2FA verification successful. Login complete."
        '400':
          description: Invalid input (e.g., invalid OTP format) or invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found."
...
  /admin/users:
    get:
      summary: List all users (Admin only)
      tags: [Admin - User Management]
      security:
        - bearerAuth: [] # Indicates admin role check needed
      responses:
        '200':
          description: A list of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError' # Admin access required
    post:
      summary: Create a new user (Admin only)
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Email address already in use.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email address already in use."
...